requirements.txt
python-dotenv~=0.15.0
django~=3.2
requests~=2.25